http-security-jersey: drop overallpting tests & dead mock scaffolding from BasicAuthFilterMisconfigurationTest
