+++ servicetalk-http-utils/src/main/java/io/servicetalk/http/utils/auth/BasicAuthHttpServiceFilter.java
@@ -72,11 +72,10 @@ public final class BasicAuthHttpServiceFilter<UserInfo> implements StreamingHttp
      * A {@link ContextMap.Key} whose presence with the value {@link Boolean#TRUE} indicates that the current request
      * was successfully authenticated by a {@link BasicAuthHttpServiceFilter}.
      * <p>
-     * This marker is stored in {@link AsyncContext} on every successful authentication, regardless of whether a
-     * {@code userInfo} key was configured via {@link Builder#userInfoAsyncContextKey(ContextMap.Key)} or
-     * {@link Builder#userInfoRequestContextKey(ContextMap.Key)}. Downstream components (for example the ServiceTalk
-     * Jersey {@code BasicAuthSecurityContextFilters}) can use it to tell a request that was authenticated but for
-     * which no user info was published apart from a request that never went through Basic authentication.
+     * Stored in {@link AsyncContext} on every successful authentication, regardless of whether a {@code userInfo} key
+     * was configured. Downstream components (for example the ServiceTalk Jersey
+     * {@code BasicAuthSecurityContextFilters}) use it to tell an authenticated request that published no user info
+     * apart from one that never went through Basic authentication.
      */
     public static final ContextMap.Key<Boolean> AUTHENTICATED =
             ContextMap.Key.newKey("io.servicetalk.http.utils.auth.basicAuthAuthenticated", Boolean.class);
@@ -395,8 +394,8 @@ public final class BasicAuthHttpServiceFilter<UserInfo> implements StreamingHttp
                                                               final StreamingHttpRequest request,
                                                               final StreamingHttpResponseFactory factory,
                                                               final UserInfo userInfo) {
-            // Always mark the request as authenticated so that downstream components can distinguish an
-            // authenticated request that carries no user info from a request that never went through Basic auth.
+            // Mark the request authenticated so downstream can tell an authenticated request with no user info from
+            // one that never went through Basic auth.
             AsyncContext.put(AUTHENTICATED, Boolean.TRUE);
             if (config.userInfoAsyncContextKey != null) {
                 AsyncContext.put(config.userInfoAsyncContextKey, userInfo);
